Refactored Config Profile and moved all Input System options into a new Input System Profile to cleanup the main config profile
