+++ Assets/MixedRealityToolkit-SDK/Profiles/DefaultMixedRealityInputSystemProfile.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ad2080e8e71c35f4e8bcde94fa68f098
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
