+++ Assets/MixedRealityToolkit/_Core/Inspectors/Profiles/MixedRealityBaseConfigurationProfileInspector.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.﻿
 
+using Microsoft.MixedReality.Toolkit.Core.Extensions.EditorClassExtensions;
 using Microsoft.MixedReality.Toolkit.Core.Managers;
 using UnityEditor;
 using UnityEngine;
@@ -9,6 +10,8 @@ namespace Microsoft.MixedReality.Toolkit.Inspectors.Profiles
 {
     public abstract class MixedRealityBaseConfigurationProfileInspector : Editor
     {
+        private static readonly GUIContent NewProfileContent = new GUIContent("+", "Create New Profile");
+
         [SerializeField]
         private Texture2D logoLightTheme = null;
 
@@ -74,6 +77,26 @@ namespace Microsoft.MixedReality.Toolkit.Inspectors.Profiles
             }
 
             return true;
+        }
+
+        protected static void RenderProfile(SerializedProperty property)
+        {
+            EditorGUILayout.BeginHorizontal();
+            EditorGUILayout.PropertyField(property);
+
+            if (property.objectReferenceValue == null)
+            {
+                if (GUILayout.Button(NewProfileContent, EditorStyles.miniButton))
+                {
+                    var profileTypeName = property.type.Replace("PPtr<$", string.Empty).Replace(">", string.Empty);
+                    Debug.Assert(profileTypeName != null, "No Type Found");
+                    ScriptableObject profile = CreateInstance(profileTypeName);
+                    profile.CreateAsset(AssetDatabase.GetAssetPath(Selection.activeObject));
+                    property.objectReferenceValue = profile;
+                }
+            }
+
+            EditorGUILayout.EndHorizontal();
         }
     }
 }
